create new preprocessing module for functions to call before running model
