+++ scripts/construct_clustered_snv_adata.py
@@ -1,11 +1,8 @@
 import click
-import pandas as pd
 import anndata as ad
-import numpy as np
 import logging
 import sys
 import pickle
-import scgenome
 import doubletime as dt
 
 
@@ -28,167 +25,50 @@ def main(adata_cna, adata_snv, tree_filename, output_cn, output_snv, output_prun
 
     tree = pickle.load(open(tree_filename, 'rb'))
 
+    # make sure the CN adata and SNV adata have the same cells
     adata = adata[snv_adata.obs.index]
+
+    # merge the sbmclone_cluster_id from the SNV adata into the CN adata
     adata.obs = adata.obs.merge(snv_adata.obs[['sbmclone_cluster_id']], left_index=True, right_index=True, how='left')
     assert not adata.obs['sbmclone_cluster_id'].isnull().any()
 
-    # perform automatic WGD depth detection if wgd_depth is negative and there is more than one clone
-    n_sbmclones = len(adata.obs['sbmclone_cluster_id'].unique())
+    # perform automatic WGD depth detection if wgd_depth is negative
     if wgd_depth < 0:
-        if n_sbmclones > 1:
-            wgd_depth = dt.tl.automatic_wgd_depth_detection(snv_adata)
-        else:
+        n_sbmclones = len(adata.obs['sbmclone_cluster_id'].unique())
+        if n_sbmclones <= 1:
             # if there is only one clone, set the WGD depth to 0
             # since the root branch is the entire tree
             wgd_depth = 0
+        else:
+            # otherwise, perform automatic WGD depth detection
+            wgd_depth = dt.pp.automatic_wgd_depth_detection(snv_adata)
         print('Automatic WGD depth detection has set wgd_depth to', wgd_depth)
 
-    # Wrangle CN anndata, identify bins with compatible cn, filter clones
-    # 
+    ### Wrangle CN anndata, identify bins with compatible cn, filter clones
 
-    # Add the modal wgd state for each sbmclone
-    adata.obs['n_wgd_mode'] = adata.obs.groupby('sbmclone_cluster_id')['n_wgd'].transform(lambda x: x.mode()[0])
-
-    # Add leaf id to copy number anndata
-    # Since multiple clones could have been combined into one leaf
-    # of the clone tree, there is not necessarily a one to one mapping
-    # of leaves to clones
-    block2leaf = {}
-    for l in tree.get_terminals():
-        for b in l.name.lstrip('clone_').split('/'):
-            block2leaf[int(b)] = l.name.lstrip('clone_') # TODO: why?
-    adata.obs['leaf_id'] = adata.obs.sbmclone_cluster_id.map(block2leaf)
-
-    # Select cells with n_wgd<=1 and n_wgd equal to the modal n_wgd
-    adata = adata[(
-        (adata.obs.n_wgd.astype(int) <= 1) &
-        adata.obs.n_wgd == adata.obs.n_wgd_mode)].copy()
-
-    # Threshold on size of clone
-    adata.obs['leaf_size'] = adata.obs.groupby('leaf_id').transform('size')
-    adata = adata[adata.obs['leaf_size'] >= min_clone_size]
-
-    # Aggregate the copy number based on the leaf id
-    adata_cn_clusters = scgenome.tl.aggregate_clusters(
-        adata,
-        agg_layers={
-            'Maj': 'median',
-            'Min': 'median',
-            'state': 'median',
-        },
-        cluster_col='leaf_id')
-    adata_cn_clusters.obs.index = adata_cn_clusters.obs.index.astype(str)
-
-    # Add per clone statistics for the frequency of the median state
-    for layer in ['state', 'Maj', 'Min']:
-        adata.layers[f'clone_median_{layer}'] = adata_cn_clusters.to_df(layer).loc[adata.obs['leaf_id'].values, :]
-        adata.layers[f'is_eq_clone_median_{layer}'] = adata.layers[layer] == adata.layers[f'clone_median_{layer}']
-        adata.var[f'is_eq_clone_median_{layer}'] = np.nanmean(adata.layers[f'is_eq_clone_median_{layer}'], axis=0)
-
-    # Redo aggregation of the copy number and include per clone stats for frequency of median state
-    adata_cn_clusters = scgenome.tl.aggregate_clusters(
-        adata,
-        agg_layers={
-            'Maj': 'median',
-            'Min': 'median',
-            'state': 'median',
-            'is_eq_clone_median_state': 'mean',
-            'is_eq_clone_median_Maj': 'mean',
-            'is_eq_clone_median_Min': 'mean',
-        },
-        agg_obs={
-            'n_wgd': 'median',
-            'haploid_depth': 'sum',
-        },
-        cluster_col='leaf_id')
-    adata_cn_clusters.obs.index = adata_cn_clusters.obs.index.astype(str)
-    adata_cn_clusters.obs['n_wgd'] = adata_cn_clusters.obs['n_wgd'].round()
-
-    # Calculate and threshold for homogeneous copy number within each clone
-    adata_cn_clusters.var['is_homogenous_cn'] = (
-        (adata_cn_clusters.var['is_eq_clone_median_Maj'] > 0.9) &
-        (adata_cn_clusters.var['is_eq_clone_median_Min'] > 0.9) &
-        (adata_cn_clusters.layers['is_eq_clone_median_Maj'] > 0.8).all(axis=0) &
-        (adata_cn_clusters.layers['is_eq_clone_median_Min'] > 0.8).all(axis=0))
-
-    # Compatible states for WGD1 and WGD0, major/minor for the snv tree model
-    compatible_cn_types = {
-        '1:0': [{'n_wgd': 1, 'Maj': 1, 'Min': 0}, {'n_wgd': 0, 'Maj': 1, 'Min': 0}],
-        '2:0': [{'n_wgd': 1, 'Maj': 2, 'Min': 0}, {'n_wgd': 0, 'Maj': 1, 'Min': 0}],
-        '1:1': [{'n_wgd': 1, 'Maj': 1, 'Min': 1}, {'n_wgd': 0, 'Maj': 1, 'Min': 1}],
-        '2:1': [{'n_wgd': 1, 'Maj': 2, 'Min': 1}, {'n_wgd': 0, 'Maj': 1, 'Min': 1}],
-        '2:2': [{'n_wgd': 1, 'Maj': 2, 'Min': 2}, {'n_wgd': 0, 'Maj': 1, 'Min': 1}],
-    }
-
-    # Check for compatibility and assign
-    adata_cn_clusters.var['snv_type'] = 'incompatible'
-    for name, cn_states in compatible_cn_types.items():
-        cn_states = pd.DataFrame(cn_states).set_index('n_wgd')
-        clone_maj = adata_cn_clusters.obs['n_wgd'].map(cn_states['Maj'])
-        clone_min = adata_cn_clusters.obs['n_wgd'].map(cn_states['Min'])
-        is_compatible = (
-            (adata_cn_clusters.layers['Maj'] == clone_maj.values[:, np.newaxis]) &
-            (adata_cn_clusters.layers['Min'] == clone_min.values[:, np.newaxis]))
-        bin_is_compatible = np.all(is_compatible, axis=0)
-        adata_cn_clusters.var.loc[bin_is_compatible, 'snv_type'] = name
-
-    # Wrangle tree, prune based on removed clusters
-    #
-
-    # Prune clones from the tree if they were removed due to size
-    remaining_leaves = adata_cn_clusters.obs.index
-    tree = scgenome.tl.prune_leaves(tree, lambda a: a.name.lstrip('clone_') not in remaining_leaves)
-    # Merge branches
-    def merge_branches(parent, child):
-        return {
-            'name': child.name,
-            'branch_length': 1,
-            'mutations': ','.join(filter(lambda a: a, [parent.mutations, child.mutations])),
-        }
-    tree = scgenome.tl.aggregate_tree_branches(tree, f_merge=merge_branches)
+    # Aggregate the CN anndata to go from cell x bin to clone x bin, filtering out clones with too few bins
+    # and SNVs that are incompatible with the doubleTime model (i.e. incorrect ASCN state or ASCN not homogenous within a clone)
+    adata_cn_clusters, tree, block2leaf = dt.pp.preprocess_cn_adata(adata, tree, min_clone_size=min_clone_size)
 
     # Manually add WGD events to the tree
     # the wgd_depth parameter controls whether a WGD event is added at the root of the tree
     # or multiple independent events are placed on branches `wgd_depth` generations below the root
-    dt.tl.add_wgd_tree(tree, adata_cn_clusters, wgd_depth=wgd_depth)
+    dt.pp.add_wgd_tree(tree, adata_cn_clusters, wgd_depth=wgd_depth)
 
     # split branches with a WGD event into two branches
-    dt.tl.split_wgd_branches(tree)
+    dt.pp.split_wgd_branches(tree)
 
     # Recursively add n_wgd to each clade
-    dt.tl.count_wgd(tree.clade, 0)
+    dt.pp.count_wgd(tree.clade, 0)
 
-    # Wrangle SNV anndata, filter based on cn adatas and merge bin information
-    #
+    ### Wrangle SNV anndata, filter based on cn adatas and merge bin information
 
-    # Filter snv adata similarly to cn adata
+    # Make sure that the SNV adata has the same cells as the CN adata
     snv_adata = snv_adata[adata.obs.index]
 
-    # Aggregate snv counts
-    snv_adata.obs['leaf_id'] = snv_adata.obs.sbmclone_cluster_id.map(block2leaf)
-    adata_clusters = scgenome.tl.aggregate_clusters(
-        snv_adata,
-        agg_layers={
-            'alt': 'sum',
-            'ref': 'sum',
-            'Maj': 'median',
-            'Min': 'median',
-            'state': 'median',
-        },
-        cluster_col='leaf_id')
-
-    # Filter snv cluster data similar to copy number
-    adata_clusters = adata_clusters[adata_cn_clusters.obs.index].copy()
-
-    # Additional layers
-    adata_clusters.layers['vaf'] = adata_clusters.layers['alt'] / (adata_clusters.layers['ref'] + adata_clusters.layers['alt'])
-    adata_clusters.layers['ref_count'] = adata_clusters.layers['ref']
-    adata_clusters.layers['alt_count'] = adata_clusters.layers['alt']
-    adata_clusters.layers['total_count'] = adata_clusters.layers['ref'] + adata_clusters.layers['alt']
-
-    # Add information from cn bin analysis
-    adata_clusters.var['snv_type'] = adata_cn_clusters.var.loc[adata_clusters.var['cn_bin'], 'snv_type'].values
-    adata_clusters.var['is_homogenous_cn'] = adata_cn_clusters.var.loc[adata_clusters.var['cn_bin'], 'is_homogenous_cn'].values
+    # Aggregate the SNV cell x snv anndata to a clone x snv anndata with the same rows and columns
+    # as adata_cn_clusters, adding the appropriate layers, obs and var for running the doubleTime model
+    adata_clusters = dt.pp.preprocess_snv_adata(snv_adata, adata_cn_clusters, block2leaf)
 
     adata_cn_clusters.write(output_cn)
     adata_clusters.write(output_snv)
